requirements.txt
fastapi
pydantic
pymysql
python-dotenv
redis
sqlalchemy
uvicorn